requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
idna==2.10
mysqlclient==1.4.6
phonenumbers==8.12.13
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
twilio==6.48.0
urllib3==1.26.2
